fixed #33
